requirements.txt
absl-py==2.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
Django==4.2.14
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.7
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.25.1
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.0.3
parso==0.8.4
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.8.2
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.1.0
zipp==3.19.2